Add go implementation of the app
This commit provides the source code and unikraft specific files
necessary to build and run a unikraft application similar to the c
and cpp ones.

The app starts a server on port 12345 and waits for a integer, responding
with the incremented value of the it.

At this point, there are some dependency errors in the build process.